requirements.txt
openpyxl==3.1.5
azure-ai-formrecognizer==3.3.3
pandas==2.2.3
python-dotenv
pandas==2.2.3
psutil==6.1.0
requests==2.32.3
openai==1.54.3
codecarbon==2.7.2
xhtml2pdf==0.2.16
jinja2==3.1.4
plotly==5.24.1
Flask
flask-socketio==5.4.1
python-socketio==5.4.1
python-engineio==4.3.4
bidict==0.23.1  
numpy==2.1.3
pytz==2024.2
python-dateutil==2.9.0.post0 
six==1.16.0
urllib3==2.2.3
idna==3.10
certifi==2024.8.30
charset-normalizer==3.4.0
typing_extensions==4.12.2
pydantic==2.9.2
annotated_types==0.7.0
pydantic_core==2.26.0
httpx==0.27.2
sniffio==1.3.1
httpcore==1.0.6
h11==0.14.0
anyio==4.6.2.post1
exceptiongroup==1.2.2
distro==1.9.0
jiter==0.7.0
pytest==8.3.3
RapidFuzz==3.10.1
py-cpuinfo==9.0.0
pynvml==11.5.3
arrow==1.3.0
prometheus-client==0.21.0
reportlab==4.2.5
pillow==11.0.0
asn1crypto==1.5.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
uritools==4.0.3
cryptography==43.0.3
cffi==1.17.1
oscrypto==1.3.0
tzlocal==5.2
qrcode==8.0
pypdf==5.1.0
packaging==24.2
html5lib
webencodings
arabic_reshaper
python-bidi==0.6.3
tqdm
